fix: relocating component
