+++ src/pages/privileges/outlets/roles/components/DetailsModal/index.tsx
